1. Fixed the conftest conflict between MCP plugin and root. 2. Re-enabled all plugins tests - all tests passing
