+++ pytest.ini
@@ -1,6 +1,6 @@
 # pytest.ini
 [pytest]
 pythonpath  = . src
-norecursedirs = skillberry-common plugins
+norecursedirs = skillberry-common
 markers =
     integration: marks tests as integration tests (require running service, deselect with '-m "not integration"')
